feature: cset-2884 refactor reportsdatabusiness
